board: rock5b-rk3588: add USB-C controller support
Enable support for the fusb302 USB Type-C controller. Since fusb302
provides the cable orientation information, this is the first step
for introducing USB-C port support for the board.

But more importantly this will do early USB PD (power deliver)
negotiation, which must happen within 5 seconds after the USB-C
connector has plugged in according to the specification. It takes
almost 5 seconds to go through the bootchain on Rock 5B and jump
to the operating system. When the Linux initializes the fusb302
usually 20-30 seconds have gone since the device has been plugged,
which is far too late. The USB PD power source reacts with a hard
reset, which disables VBUS for some time. This is not a problem for
a battery driven device, but Rock 5B will loose its power-supply
and reset. By initializing PD in U-Boot, this can be avoided.

Signed-off-by: Sebastian Reichel <sebastian.reichel@collabora.com>